Adds NewsletterLive
We add NewsletterLive: a live view to test form changes.

The form is a sign up form to subscribe to a newsletter.

We don't test that the form submission works. Instead, we test that the
form change with invalid data results in an error.

Because email validation via regex is notoriously difficult to get
right, we do the simplest of validations -- it must have an `@` sign.

We also add a second test to show that the `form/3` helper continues to
be very valuable. In the case of using `render_change/3` (targeting the
event directly), we don't get any feedback that the `phx-change`
attribute isn't set.

Test files
----------

As usual, we include two test files:

- An empty `newsletter_live_test.exs` so we can add tests, and
- A `newsletter_live_done_test.exs` that contains the two tests in their
  completed state.