Adds navigation between live views
What changed?
============

We add navigation between LiveViews. To test that, we introduce a
`DashboardLive` and `TeamLive`. The two live views are under the same
`live_session` so we can navigate with `<.link navigate={}>`.

We also test navigation that requires a full page reload (such as
navigating to another page or to a live view in a different
`live_session`). That navigation which uses `<.link href={}>`.

The testing is all done from the "dashboard" perspective, so we
introduce a `dashboard_live_test.exs` and the equivalent "done" version.