Adds AlbumLive.Index and Show
What changed?
============

We introduce two new live views: `AlbumLive.Index` and `AlbumLive.Show`.

Through the index page we can create a new album with name and upload a
series of photos. The aim of the commit is to learn how to test uploads.

All the photos are stored in the server in the `/uploads` static
directory.

To do that, we add "uploads" to the list of static paths.

We also define a few top-level `Ranger` functions for uploads paths
that look for a configured upload path or default to
`priv/static/uploads`.

That allows our development uploads to live in `priv/static/uploads/`
but our test uploads to be configured and live in
`priv/static/uploads/test/`. That's a nice quality of life improvement
since we clean up uploads during tests by removing the directory.

As usual, we include two test files: `album_live/index_test.exs` and the
"done" counterpart.