Add MovieLive to test external uploads
What changed?
============

We add `MovieLive` to test external uploads.

We do something very similar to `AlbumLive` but we add one new test and
modify another one:

1. We add a test that ensures we create the correct metadata for
   external uploads.

2. We modify the "submit" test to ensure we're storing the correct URLs
   in the movie record

As usual, we have the normal test and the "done" equivalent. But in this
case, the normal test has one test that we'll reuse in the second point
mentioned above.